// src/types/ProfileTypes.ts
export enum PROFILES {
    BASIC = 'urn:cen.eu:en16931:2017#compliant#urn:factur-x.eu:1p0:basic',
    BASIC_WITHOUT_LINES = 'urn:factur-x.eu:1p0:basicwl',
    MINIMUM = 'urn:factur-x.eu:1p0:minimum',
    COMFORT = 'urn:cen.eu:en16931:2017',
    EXTENDED = 'urn:cen.eu:en16931:2017#conformant#urn:factur-x.eu:1p0:extended'
}